requirements.txt
claude-code-sdk>=0.0.14
PyYAML>=6.0
schedule>=1.2.0
python-dateutil>=2.8.0
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-mock>=3.10.0